refactor: moved component and changed names
